Simplify lock window creation
Collapse lock window creation onto the tracked helper and normalize the project-owned tracked-window count boundary to size_t. Keep the Xlib-owned unmap counts unchanged while preserving the existing lock window cleanup and failure behavior.

+++ lock_state.h
@@ -55,7 +55,7 @@ struct LockBlankingState {
 
 struct LockWindows {
   Window tracked_windows[LOCK_TRACKED_WINDOW_CAPACITY];
-  unsigned int tracked_window_count;
+  size_t tracked_window_count;
   Window root_window;
   Window parent_window;
   Window background_window;
